requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
pillow>=9.0.0
jupyter>=1.0.0
opencv-python>=4.6.0
tensorflow>=2.10.0  # For CNN models (retinal/product images)
torch>=1.13.0  # Alternative to TensorFlow
streamlit>=1.35.0
scikeras>=0.13.0
pyyaml>=6.0.1
#tensorflow-cpu==2.15.1
#keras==2.15.0
#h5py==3.11.0
#opencv-python-headless==4.10.0.84
#numpy==1.26.4
#pandas==2.2.2
#streamlit==1.38.0
gdown==4.7.1
#Pillow==10.4.0
#scikit-learn==1.5.1
#matplotlib>=3.5.0
#seaborn>=0.12.0
#pillow>=9.0.0
#jupyter>=1.0.0
#torch>=1.13.0  # Alternative to TensorFlow
#pyyaml>=6.0.1
# Google Drive API (googleapiclient) + auth stack
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-httplib2>=0.2.0
httplib2>=0.22.0
# Public-link downloader path uses this (likely already present)
requests>=2.28.0